pyproject.toml
[project]
name = "pridexyz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colorlog>=6.10.1",
    "colour-science>=0.4.7",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.21.1",
    "urllib3>=2.6.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
